+++ client/src/App.tsx
@@ -1,54 +1,15 @@
 import React from 'react';
 import { BrowserRouter as Router, Switch } from "react-router-dom";
-import Landing from './views/Landing/landing';
-import Find from './views/Find/Find';
-import About from './views/About/About';
-import TheTeam from './views/TheTeam/TheTeam';
-import TempProfile from './views/CreateProfile/Profile';
-import UserProfile from './views/UserProfile/UserProfile'
-import CreateProfile from './views/CreateProfile/CreateProfile';
-import Footer from './components/Footer/Footer';
 import Loading from './components/Loading/Loading';
 import { useAuth0 } from '@auth0/auth0-react';
-import ProtectedRoute from './auth/protected-route';
 import Navbar from './components/Navbar/Navbar';
-import { Container } from "@material-ui/core";
 import { makeStyles, createStyles, Theme } from "@material-ui/core/styles";
+import LoggedInLanding from "./views/Landing/loggedInLanding"
+import LoggedOutLanding from "./views/Landing/loggedOutLanding"
 
-const useStyles = makeStyles((theme: Theme) => createStyles({
-  root: {
-    display: `flex`,
-    backgroundSize: "cover",
-    backgroundRepeat: 'no-repeat',
-    backgroundPosition: 'center',
-    height: "100vh",
-    justifyContent: `center`
-  }
-}));
-
-
-const App: React.FC = (props: any): JSX.Element => {
+const App: React.FC = (): JSX.Element => {
   
-  const { isAuthenticated } = useAuth0();
-  let bgArray: Array<number> = [1, 2, 3, 4]
-
-  const arrayShuf = (array: Array<number>): Array<number> => {
-    let j: number = 0;
-    let temp: number;
-
-    for (let i = array.length - 1; i >= array.length - 3; i--) {
-      j = Math.floor(Math.random() * (i + 1));
-      temp = array[i];
-      array[i] = array[j];
-      array[j] = temp;
-    }
-
-    return array;
-  }
-
-  const shufArray: Array<number> = arrayShuf(bgArray);
-  const classes = useStyles();
-  const { isLoading } = useAuth0();
+  const { isAuthenticated, isLoading } = useAuth0();
 
   if (isLoading) {
     return <Loading />;
@@ -58,22 +19,7 @@ const App: React.FC = (props: any): JSX.Element => {
     <Router>
         <>
           <Navbar />
-          <Switch>
-            <ProtectedRoute exact path="/tempprofile" component={TempProfile} />
-            <ProtectedRoute exact path="/createprofile" component={CreateProfile} />
-            <ProtectedRoute exact path="/userProfile" component={UserProfile} />
-            <ProtectedRoute exact path="/find" component={Find} />
-          </Switch>
-          <Container id="landing" className={classes.root} maxWidth='xl' style={{ backgroundImage: `url(/backgrounds/background${shufArray[0]}.jpg)` }} >
-            <Landing />
-          </Container>
-          <Container id="about" className={classes.root} maxWidth='xl' style={{ backgroundImage: `url(/backgrounds/background${shufArray[1]}.jpg)` }} >
-            <About />
-          </Container>
-          <Container id="theteam" className={classes.root} maxWidth='xl' style={{ backgroundImage: `url(/backgrounds/background${shufArray[2]}.jpg)` }} >
-            <TheTeam />
-          </Container>
-          {/* <Footer /> */}
+          { isAuthenticated ? <LoggedInLanding /> : <LoggedOutLanding /> }
         </>
     </Router>
