+++ client/src/components/Landing/landing.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { useTheme } from '@material-ui/core/styles';
 import useMediaQuery from '@material-ui/core/useMediaQuery';
-import AuthNav from "../../components/Login/Auth-nav";
-import SignupButton from "../../components/Login/Signup-button"
+import AuthNav from "../Login/Auth-nav";
+import SignupButton from "../Login/Signup-button"
 
 import { Box, Container } from "@material-ui/core"
 import { makeStyles, createStyles, Theme } from "@material-ui/core/styles";
