+++ client/src/components/TheTeam/TheTeam.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useTheme } from '@material-ui/core/styles';
 import { Box, Grid, Paper, Container, Avatar, Typography } from "@material-ui/core"
 import { makeStyles, createStyles, Theme } from "@material-ui/core/styles";
-import TeamBio from '../../components/TeamBio/TeamBio'
+import TeamBio from '../TeamBio/TeamBio'
 
 const useStyles = makeStyles((theme: Theme) => createStyles({
     root: {
