+++ client/src/components/UserProfile/UserProfile.tsx
@@ -10,60 +10,59 @@ import { RootState } from "../../redux/store";
 import API from "../../utils/API";
 
 export interface IUser {
-    nickName: string,
-    firstName: string,
-    lastName: string,
-    intentionStatus: string,
-    location: string,
-    email: string,
-    phone: string,
-    blurb: string,
+  nickName: string,
+  firstName: string,
+  lastName: string,
+  intentionStatus: string,
+  location: string,
+  email: string,
+  phone: string,
+  blurb: string,
 }
 
 const UserProfile = () => {
   const userProfile: any = useAuth0();
 
-//  console.log(userProfile.user)
+  //  console.log(userProfile.user)
 
   const [user, setUser] = useState<IUser[]>([])
 
   const writeProfile = (res: any[]) => {
     const users: IUser[] = res.map((user) => {
-        return {
-            nickName: user.nickName,
-            firstName: user.firstName,
-            lastName: user.lastName,
-            intentionStatus: user.intentionStatus,
-            location: user.location,
-            email: user.email,
-            phone: user.phone,
-            blurb: user.blurb
-        }
+      return {
+        nickName: user.nickName,
+        firstName: user.firstName,
+        lastName: user.lastName,
+        intentionStatus: user.intentionStatus,
+        location: user.location,
+        email: user.email,
+        phone: user.phone,
+        blurb: user.blurb
+      }
     })
     console.log(users)
     setUser(users)
-    
-}
 
+  }
 
-useEffect(() => {
+  useEffect(() => {
     API.getUsers().then(res => {
-        const findUser = res.data;
-        const thisUser = findUser.filter((findUser: any) => userProfile.user.email === findUser.email)
+      const findUser = res.data;
+      const thisUser = findUser.filter((findUser: any) => userProfile.user.email === findUser.email)
 
-        // console.log(thisUser[0])
-        writeProfile(thisUser)
-        // dbUsers(res.data);
+      // console.log(thisUser[0])
+      writeProfile(thisUser)
+      // dbUsers(res.data);
     })
-}, [])
+  }, [])
 
   return (
     <div className="paddingfix">
-{user.map((user) => {
-    return (
-        <Container  maxWidth="lg" >
+      {user.map((user) => {
+        return (
+          <Container maxWidth="lg" >
             <h1>User Profile Page</h1>
-    
+
             <h1 key={user.nickName}>{user.nickName}</h1>
             <p key={user.firstName}>{user.firstName} {user.lastName}</p>
             <p key={user.intentionStatus}>{user.intentionStatus}</p>
@@ -71,11 +70,10 @@ useEffect(() => {
             <p key={user.email}>{user.email}</p>
             <p key={user.phone}>{user.phone}</p>
             <p key={user.blurb}>{user.blurb}</p>
-
-    </Container>
-)
-})}
-</div>
+          </Container>
+        )
+      })}
+    </div>
   );
 };
